+++ subject11/my-app/.storybook/main.js
@@ -1,7 +1,8 @@
 module.exports = {
   "stories": [
     "../src/**/*.stories.mdx",
-    "../src/**/*.stories.@(js|jsx|ts|tsx)"
+    "../src/**/*.stories.@(js|jsx|ts|tsx)",
+    "../src/stories/my-app/*.stories.js"
   ],
   "addons": [
     "@storybook/addon-links",
