+++ subject11/my-app/src/components/Square.jsx
@@ -1,0 +1,9 @@
+import React from "react"
+
+export function Square(props) {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
