+++ subject11/my-app/src/stories/my-app/Board.stories.js
@@ -1,4 +1,4 @@
-import { Board } from "../..";
+import { Board } from "../../components/Board"
 
 export default {
   title: 'Board Component',
