+++ subject11/my-app/src/stories/my-app/Game.stories.js
@@ -1,4 +1,4 @@
-import { Game } from "../..";
+import { Game } from "../../components/Game"
 
 export default {
   title: 'Game Component',
