+++ subject11/my-app/src/stories/snapshot.test.js
@@ -1,0 +1,7 @@
+import initStoryshots from '@storybook/addon-storyshots';
+
+import { imageSnapshot } from '@storybook/addon-storyshots-puppeteer';
+
+initStoryshots({
+  test: imageSnapshot({})
+})
